(#10) Removed all warnings, mostly by using disables for dependencies, and using size_t instead of uint32_t for counting variables

+++ Ramensoup/src/Platform/OpenGL/OpenGLIndexBuffer.cpp
@@ -6,7 +6,7 @@
 namespace Ramensoup
 {
 
-	OpenGLIndexBuffer::OpenGLIndexBuffer(const uint32_t* verticies, uint32_t count)
+	OpenGLIndexBuffer::OpenGLIndexBuffer(const uint32_t* verticies, size_t count)
 		:m_Count(count)
 	{
 		glGenBuffers(1, &m_RendererID);
