+++ Ramensoup/src/Platform/OpenGL/OpenGLVertexBuffer.cpp
@@ -12,7 +12,7 @@ namespace Ramensoup
 
 		glBufferData(GL_ARRAY_BUFFER, size, nullptr, GL_DYNAMIC_DRAW);
 	}
-	OpenGLVertexBuffer::OpenGLVertexBuffer(const void* data, uint32_t size)
+	OpenGLVertexBuffer::OpenGLVertexBuffer(const void* data, size_t size)
 	{
 		glGenBuffers(1, &m_RendererID);
 		glBindBuffer(GL_ARRAY_BUFFER, m_RendererID);
